Keep only Docker image related changes
